+++ plugins/core/main.py
@@ -15,9 +15,10 @@ app.add_dir_for_search("plugins/core/actions", include_private=False)
 app.update_config(
     {
         "core": {
-            "music-download": False,
+            "music-download": True,
             "usb-default": ["linux foundation", "webcam", "network", "finger"]
         }
     }
 )
+