+++ plugins/custom/actions/tiktok.py
@@ -7,10 +7,11 @@ import lxml
 
 from api import app
 from data.constants import PLUGINS_DIR
+from utils import run
 
 
 def move():
-    WINDOW_KEYWORD = "ULTRAKILL"  # Part of window title or class
+    WINDOW_KEYWORD = "Zen Browser"  # Part of window title or class
     MONITOR_POSITIONS = [(0, 0), (1920, 0)]  # Example: two monitors, second starts at x=1920
     TARGET_MONITOR = 1  # Index of target monitor (0-based)
 
@@ -37,7 +38,7 @@ def move():
             x, y = MONITOR_POSITIONS[TARGET_MONITOR]
             move_window(win_id, x, y)
             break
-        time.sleep(0.5)
+        time.sleep(0.25)
 
 
 def ultrakill(**kwargs):
@@ -47,7 +48,7 @@ def ultrakill(**kwargs):
         "set",
         "org.gnome.desktop.background",
         "picture-uri-dark",
-        "file:///home/ilyamiro/Изображения/wallpaper_black.jpg"
+        "file:///home/ilyamiro/Изображения/Wallpapers/wallpaper_black.jpg"
     ])
     subprocess.run(
         ["sudo", "tee", "/sys/class/leds/asus::kbd_backlight/brightness"],
@@ -61,7 +62,7 @@ def ultrakill(**kwargs):
 
     subprocess.Popen(["alacritty", "-e", "cava"])
 
-    app.audio.play(f"{PLUGINS_DIR}/custom/actions/holdme.mp3")
+    app.audio.play(f"{PLUGINS_DIR}/custom/actions/ultra.mp3")
 
     move()
 
@@ -70,7 +71,7 @@ app.add_func_for_search(ultrakill)
 
 app.manager.add(
     app.Command(
-        keywords=["прошлый", "раз"],
+        keywords=["кого", "я", "обманываю"],
         action="ultrakill",
         tts=True,
         responses=[]
@@ -104,3 +105,42 @@ app.manager.add(
         }
     )
 )
+
+
+def home1():
+    app.say("С возвращением, сэр! Добро пожаловать домой. Хотите ли вы чтобы я включил музыку?")
+    subprocess.run(
+        ["sudo", "tee", "/sys/class/leds/asus::kbd_backlight/brightness"],
+        input="3\n",
+        text=True
+    )
+    subprocess.run(["wmctrl", "-k", "on"])
+    subprocess.run(["brightnessctl", "set", "100%"])
+    subprocess.Popen(["bash", "/home/ilyamiro/Документы/pycharm-community-2023.2.3/bin/pycharm.sh"])
+    subprocess.Popen(["xdg-open", "https://ilyamiro.github.io/stewart/docs/Getting%20Started"])
+    move()
+
+
+def home2():
+    link = "https://fluxfm.streamabc.net/flx-chillhop-mp3-128-8581707?sABC=6825027p%230%233n1o17opo5312o497o69212842qq72s8%23fgernzf.syhksz.qr&aw_0_1st.playerid=streams.fluxfm.de&amsparams=playerid:streams.fluxfm.de;skey:1747255932"
+    app.audio.stream(link)
+
+
+home_scenario = app.Scenario("home", app.Timeline([
+    [
+        app.Trigger(
+            ["я", "дома"],
+            home1,
+        )
+    ],
+    [
+        app.Trigger(
+            ["давай"],
+            home2
+
+        )
+    ]
+]), max_gap=3)
+app.add_scenario(home_scenario)
+
+